requirements.txt
torch 
torch_geometric 
sentence-transformers
google-genai
pandas
tabulate
pymilvus
milvus-lite